requirements.txt
# Core Backend
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.6.1
python-multipart==0.0.6

# Data Processing
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.3

# Auth & Security
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7  # For encryption

# Frontend
streamlit==1.29.0
requests==2.31.0

# Code Protection & Packaging
pyarmor==8.4.2         # For code obfuscation
pyinstaller==6.2.0      # For creating executables
cython==3.0.6          # For compiling to C
nuitka==1.9.0          # Alternative compiler

# Development & Utilities
python-dateutil==2.8.2
jupyterlab==4.0.10     # Optional for development
black==23.11.0         # Code formatting
pytest==7.4.3          # Testing

# Performance
gunicorn==21.2.0       # Production server
uvloop==0.19.0         # Faster ASGI server (optional)

# Database (if needed)
#sqlalchemy==2.0.23
#psycopg2-binary==2.9.9  # PostgreSQL adapter